Consolidate duplicate update logic into CliUpdateHelper
Extract shared utilities from UpdateCommand, AutoUpdater, and CliDownloader
into a single CliUpdateHelper static class:

- Platform detection (OS + architecture)
- Download with timeout + SHA-512 checksum validation
- Executable replacement with backup-and-swap pattern
- Old backup file cleanup
- IsRunningAsDotNetTool check
- Download URL construction

Removes ~360 lines of duplicated code across the three files.
