+++ tests/Aspire.Cli.Tests/Commands/UpdateCommandTests.cs
@@ -94,7 +94,7 @@ public class UpdateCommandTests(ITestOutputHelper outputHelper)
         var updateCommand = CreateUpdateCommand(workspace);
 
         // Act
-        updateCommand.CleanupOldBackupFiles(targetExePath);
+        UpdateCommand.CleanupOldBackupFiles(targetExePath);
 
         // Assert
         Assert.False(File.Exists(oldBackup1), "Old backup file should be deleted");
@@ -117,7 +117,7 @@ public class UpdateCommandTests(ITestOutputHelper outputHelper)
         var updateCommand = CreateUpdateCommand(workspace);
 
         // Act & Assert - should not throw exception
-        updateCommand.CleanupOldBackupFiles(targetExePath);
+        UpdateCommand.CleanupOldBackupFiles(targetExePath);
 
         // On Windows, locked files cannot be deleted, so the file should still exist
         // On Mac/Linux, locked files can be deleted, so the file may be deleted
@@ -140,7 +140,7 @@ public class UpdateCommandTests(ITestOutputHelper outputHelper)
         var updateCommand = CreateUpdateCommand(workspace);
 
         // Act & Assert - should not throw exception
-        updateCommand.CleanupOldBackupFiles(nonExistentPath);
+        UpdateCommand.CleanupOldBackupFiles(nonExistentPath);
     }
 
     [Fact]
@@ -152,7 +152,7 @@ public class UpdateCommandTests(ITestOutputHelper outputHelper)
         var updateCommand = CreateUpdateCommand(workspace);
 
         // Act & Assert - should not throw exception
-        updateCommand.CleanupOldBackupFiles(targetExePath);
+        UpdateCommand.CleanupOldBackupFiles(targetExePath);
     }
 
     private UpdateCommand CreateUpdateCommand(TemporaryWorkspace workspace)
